stlink: Adjusted config to use System-Reset only.
This to make sure that the linker is able to reset the target
before connection, else it might end up in unsuccessful connections.

And also, adjusted the programming command so that configurations
are set in the config file and break up long commands into multi-line
commands in order to fit the 80-chars boundary.